Move CreatingCard.
This is to be shared between create flow and connect flow later.
